TcasesCommand: Refactor using TcasesIO
