+++ tcases-cli/src/main/java/org/cornutum/tcases/TcasesCommand.java
@@ -30,7 +30,7 @@ import java.util.Properties;
  * Generates a set of {@link TestCase test cases} from a {@link SystemInputDef system input definition}.
  *
  */
-public class TcasesCommand extends Tcases
+public class TcasesCommand
   {
   /**
    * Represents a set of command line options.
@@ -981,8 +981,9 @@ public class TcasesCommand extends Tcases
   /**
    * Creates a new Tcases object.
    */
-  public TcasesCommand()
+  private TcasesCommand()
     {
+    // Static methods only
     }
 
   /**
@@ -1283,53 +1284,6 @@ public class TcasesCommand extends Tcases
           genWriter.close();
           }
         }
-      }
-    }
-
-  /**
-   * Returns test case definitions for the given system input definition, using the given generator set and
-   * base test definitions. If <CODE>genDef</CODE> is null, the default generator is used.
-   * If <CODE>baseDef</CODE> is null, no base test definitions are used.
-   */
-  public static SystemTestDef getTests( InputStream inputDefStream, InputStream genDefStream, InputStream baseDefStream)
-    {
-    try
-      {
-      return
-        getTests
-        ( new SystemInputDocReader( inputDefStream).getSystemInputDef(),
-          genDefStream==null? null : new GeneratorSetDocReader( genDefStream).getGeneratorSet(),
-          baseDefStream==null? null : new SystemTestDocReader( baseDefStream).getSystemTestDef());
-      }
-    catch( Exception e)
-      {
-      throw new RuntimeException( "Can't get test definitions", e);
-      }
-    }
-
-  /**
-   * Returns new test case definitions for the given system input definition, using the default generator.
-   */
-  public static SystemTestDef getTests( InputStream inputDefStream)
-    {
-    return getTests( inputDefStream, null, null);
-    }
-
-  /**
-   * Writes an XML document describing given test case definitions to the given output stream.
-   */
-  @SuppressWarnings("resource")
-  public static void writeTests( SystemTestDef testDef, OutputStream outputStream)
-    {
-    try
-      {
-      SystemTestDocWriter writer = new SystemTestDocWriter( outputStream);
-      writer.write( testDef);
-      writer.flush();
-      }
-    catch( Exception e)
-      {
-      throw new RuntimeException( "Can't write test definitions", e);
       }
     }
